pyproject.toml
[project]
name = "dl-toolkit"
version = "0.1.0"
description = "Personal DL toolkit to build effective models"
readme = "README.md"
authors = [
    { name = "Vlad Sorokin", email = "neuromancer.ai.lover@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "gitpython>=3.1.44",
    "natsort>=8.4.0",
    "pytweening>=1.2.0",
    "semver>=3.0.4",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.uv]
no-build-isolation-package = ["torch", "torchvision"]

[dependency-groups]
dev = [
    "jupyterlab>=4.3.5",
    "matplotlib>=3.10.0",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
]


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
